Updated documentation and added new functions

+++ include/time_shield_cpp/parts/time_struct.hpp
@@ -8,14 +8,16 @@
 
 namespace time_shield {
 
+    /// \ingroup time_structures
     /// \brief Structure to represent time.
     struct TimeStruct {
-        int hour;   ///< Hour component of time (0-23)
-        int min;    ///< Minute component of time (0-59)
-        int sec;    ///< Second component of time (0-59)
-        int ms;     ///< Millisecond component of time (0-999)
+        int16_t hour;   ///< Hour component of time (0-23)
+        int16_t min;    ///< Minute component of time (0-59)
+        int16_t sec;    ///< Second component of time (0-59)
+        int16_t ms;     ///< Millisecond component of time (0-999)
     };
 
+    /// \ingroup time_structures
     /// \brief Creates a TimeStruct instance.
     /// \param hour The hour component of the time.
     /// \param min The minute component of the time.
@@ -23,10 +25,10 @@ namespace time_shield {
     /// \param ms The millisecond component of the time, defaults to 0.
     /// \return A TimeStruct instance with the provided time components.
     inline const TimeStruct create_time_struct(
-            int hour,
-            int min,
-            int sec = 0,
-            int ms = 0) {
+            int16_t hour,
+            int16_t min,
+            int16_t sec = 0,
+            int16_t ms = 0) {
         TimeStruct data{hour, min, sec, ms};
         return data;
     }
